Verify installer script checksum when bootstrapping
We verify the downloaded install script's SHA-256 value against a known
checksum in bin/hermit. This ensures the integrity of the installer
when bootstrapping hermit.

We version (via file name suffix) the install script by the script's
short hash value, so that an install script can be uniquely identified
by its version / file name. The corresponding bin/hermit is generated
with hermit init for a project.

Note that an install script's "version" is unrelated to the version of
the hermit executable.